+++ conf/genmon.conf
@@ -12,6 +12,8 @@ port = /dev/serial0
 # a TCP socket connection, it will not setup the port. The software expects
 # the TCP connection to be a passthrough (i.e. not process modbus protocol).
 # The project software will process the modbus protocol.
+# If modbus_tcp is enabled the inststead of serial passthrough over TCP, the
+# modbus TCP protocol will be used.
 use_serial_tcp = False
 
 # if use_serial_tcp is enabled, this parameter sets the IP address used.
@@ -19,6 +21,11 @@ serial_tcp_address =
 
 # if use_serial_tcp is enabled, this parameter sets the IP port used.
 serial_tcp_port = 8899
+
+# enable to use Modbus TCP protocol over TCP/IP. Note serial_tcp_address and
+# serial_tcp_port must be set. Typically serial_tcp_port is set to 502 to be
+# compiant with modbus tcp standards
+modbus_tcp = False
 
 # If this option is set to true, weekly exercise time and remote start / stop
 # commands will be handled by the transfer switch and not the generator
