+++ data/tooltips.txt
@@ -77,6 +77,7 @@ update_check = If enabled the software will send an email notice when a software
 use_serial_tcp = If enabled, the software will use TCP over serial instead of direct serial communications
 serial_tcp_address = If Serial over TCP/IP is enabled, this parameter sets the IP address used for the serial server.
 serial_tcp_port = If Serial over TCP/IP is enabled, this parameter sets the IP port used for the serial server.
+modbus_tcp = If enabled, the modbus TCP protocol will be used. If disabled modbus RTU (serial) over TCP is used.
 hts_transfer_switch=Enable if you have a HTS, MTS or STS transfer switch to enable reading additional data from the transfer switch
 smart_transfer_switch = If enabled a smart transfer switch is utilized and the generator controller will not be responsible for starting and stopping or or weekly exercise functions and these items will be disabled in the web interface.
 serialnumberifmissing= If the serial number is not present due to the controller being replaced you can override the missing serial number and still have the software lookup model information. This is a numeric value that must be ten digits. Pad leading zeros if you serial number is less than 10 digits.
@@ -310,3 +311,176 @@ usecalculatedpower=Calculate power output from power factor, average current and
 0ebc=External Switch Backup Battery Voltage
 0ebf=External Switch SW Version
 0ec0=External Switch Selected
+
+[powerzone]
+0000 = RA Status 0
+0001 = RA Status 1
+0002 = RA Status 2
+0003 = RA Status 3
+0004 = RA Status 4
+0005 = RA Status 5
+0006 = RA Status 6
+0007 = RA Status 7
+0008 = RA Status 8
+0009 = RA Status 9
+006a = DI Results 0
+006B = DI Results 0
+0070 = DOF Results 0
+0071 = DOF Results 1
+0072 = DOF Results 2
+015d = Gen Time Hr Min
+015e = Gen Time Sec DyWk
+015f = Gen Time Month Day
+0160 = Gen Time Yr
+0161 = Alarm Global CMD
+0162 = Alarm Globals
+0163 = Engine Hour Now
+0165 = Engine Status
+0166 = Generator Status
+0171 = Gen Current Phase Rotation
+0173 = Gen Voltage Phase Rotation
+019a = Quite Test Request
+019b = Engine KW Hours
+01a2 = O2 Sensor
+01a4 = Governor Position
+01b5 = Alarm Global Flags
+01b6 = Key Switch State
+0167 = Maint Life Remaining
+01b7 = Battery Charger Current
+01b8 = Battery Voltage
+01b9 = Bus Average Current
+01bb = Bus Average Voltage LL
+01c3 = Bus N Current
+01c5 = Bus Frequency
+01cd = Bus Volt Phase A B
+01ce = Bus Volt Phase A N
+01cf = Bus Volt Phase B C
+01d0 = Bus Volt Phase B N
+01d1 = Bus Volt Phase C A
+01d2 = Bus Volt Phase C N
+01d3 = Coolant Level
+01d4 = Coolant Temp
+01d5 = Gen Phase A Current
+01d7 = Gen Phase B Current
+01d9 = Gen Phase C Current
+01dd = DEF Level
+01e0 = Fuel Level
+01e1 = Fuel Pressure
+01e2 = Gen Average Current
+01e4 = Gen Average Voltage LL
+01e5 = Gen Average Voltage LN
+01e6 = Output Frequency
+01e7 = Turbo Boost Pressure
+01e8 = Intake Air Temperature
+01e9 = Gen Total KVA
+01eb = Gen Total KVAR
+01ed = KW Hours
+01ef = Gen Last KW Hr
+01f1 = Last Run Hours
+01f2 = Oil Level
+01f3 = Oil Pressure
+01f4 = Oil Temp
+01fe = Intake Manifold Air Temperature
+0203 = Total Power KW
+0205 = Gen Total Power Factor
+0206 = Output RPM
+020a = Engine Hours Now
+020c = System Status
+020d = Gen Volt Phase A B
+020e = Gen Volt Phase A N
+020f = Gen Volt Phase B C
+0210 = Gen Volt Phase B N
+0211 = Gen Volt Phase C A
+0212 = Gen Volt Phase C N
+0214 = Battery Temperature
+0215 = Sync Angle
+0216 = MPS Capacity
+0218 = Alarm LOG IDX
+0219 = Event LOG IDX
+02c1 = Switch Closed
+02c4 = Remote Start
+007d = Gen Voltage AN
+007f = Gen Voltage BN
+0081 = Gen Voltage CN
+0083 = Gen Voltage LN Avg
+0085 = Bus Voltage AN
+0087 = Bus Voltage BN
+0089 = Bus Voltage CN
+008B = Bus Voltage LN Avg
+008d = Gen Voltage AB
+008f = Gen Voltage BC
+0091 = Gen Voltage CA
+0093 = Gen Voltage LL Avg
+0095 = Bus Voltage AB
+0097 = Bus Voltage BC
+0099 = Bus Voltage CA
+009b = Bus Voltage LL Avg
+009d = Gen Current A
+009f = Gen Current B
+00a1 = Gen Current C
+00a3 = Gen Current Avg
+00a5 = Gen Current N
+00a7 = Bus Current A
+00a9 = Bus Current B
+00ab = Bus Current C
+00ad = Bus Current Avg
+00af = Bus Current N
+00b1 = Gen Apparent Pwr LL Total
+00b3 = Gen Real Pwr KW LL Total
+00b5 = Gen Real Pwr Watts LL Total
+00b7 = Gen Reactive Pwr LL Total
+00b9 = Gen PF LL Avg
+00bb = Bus Apparent Pwr LL Total
+00bd = Bus Real Pwr KW LL Total
+00bf = Bus Real Pwr Watts LL Total
+00c1 = Bus Reactive Pwr LL Total
+00c3 = Bus PF LL Avg
+00c5 = AVR Field Current AVR SCR Delay Angle
+00c7 = AVR Excitation Voltage Frequency
+00c9 = Gen Freq LL
+00cb = Bus Freq LL
+00cd = RPM 1
+00cf = Batt Voltage
+00d1 = Batt Charger Current 1
+00d3 = Oil Temp 1
+00d5 = Oil Press 1
+00d7 = Oil Press 2
+00d9 = Coolant Temp 1
+00db = Coolant Temp 2
+00dd = Coolant Level Engine 1
+00df = Governor Throttle Position 1
+00e1 = Air Fuel Ratio Duty Cycle 1
+00e3 = Pre CAT O2 Sensor 1
+00e5 = Intake Manifold Air Temp 1
+00e7 = AVR Voltage Bias
+00e9 = Gov Speed Bias
+00eb = Engine Percent Load at Speed
+00ed = Engine Desired Speed
+00ef = Boost Press Post Turbo 1
+00f1 = Engine Fuel Delivery Pressure
+00f3 = Diesel Fuel Temp
+00f5 = Diesel Fuel Flow Rate
+00f7 = Diesel Fuel Level 1
+00f9 = Mixer Position
+00fb = Intake Air Pressure Delta Filter
+00fd = Air Intake Pressure
+0241=External Switch General Status
+0242=Ext Switch Mimic Diagram
+0247=External Switch Target Voltage
+024a=External Switch Target Freq
+024d=External Switch Utility AB Voltage
+024e=External Switch Utility BC Voltage
+024f=External Switch Utility CA Voltage
+0250=External Switch Utility A Amps
+0251=External Switch Utility B Amps
+0252=External Switch Utility C Amps
+0254=External Switch Average Utility Volts
+0255=External Switch Average Utility Amps
+0256=External Switch Utility Freq
+0257=External Switch Utility Power Factor
+0258=External Switch Utility Power
+0259=External Switch Generator Average Voltage
+025a=External Switch Generator Freq
+025c=External Switch Backup Battery Voltage
+025f=External Switch SW Version
+0260=External Switch Selected
